lots of refactoring
