+++ jobs/furnishings/src/furnishings/config.py
@@ -18,6 +18,7 @@ import sys
 
 from dotenv import find_dotenv, load_dotenv
 
+
 load_dotenv(find_dotenv())
 
 CONFIGURATION = {
@@ -41,7 +42,7 @@ def get_named_config(config_name: str = 'production'):
     return config
 
 
-class _Config(object):  # pylint: disable=too-few-public-methods
+class _Config:  # pylint: disable=too-few-public-methods
     """Base class configuration."""
 
     PROJECT_ROOT = os.path.abspath(os.path.dirname(__file__))
