+++ jobs/furnishings/src/furnishings/utils/__init__.py
@@ -1,24 +1,14 @@
-# Copyright © 2021 Province of British Columbia.
+# Copyright © 2024 Province of British Columbia
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#    http://www.apache.org/licenses/LICENSE-2.0
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Installer and setup for this module."""
-from glob import glob
-from os.path import basename, splitext
-from setuptools import find_packages, setup
-
-setup(
-    name='furnishings',
-    packages=find_packages('src'),
-    package_dir={'': 'src'},
-    py_modules=[splitext(basename(path))[0] for path in glob('src/*.py')]
-)
+"""This module holds general utility functions and helpers for the main package."""
