+++ jobs/furnishings/tests/unit/stage_processors/test_stage_one.py
@@ -21,36 +21,18 @@ from datetime import datetime
 from unittest.mock import MagicMock, patch
 
 import pytest
-from legal_api.models import Business, Configuration, Furnishing
+import requests
+from legal_api.models import Business, Furnishing
+from legal_api.services.bootstrap import AccountService
 from registry_schemas.example_data import FILING_HEADER, RESTORATION
 
-from furnish import check_run_schedule, get_email_address_from_auth, stage_1_process
+from furnishings.stage_processors.stage_one import StageOneProcessHelper, process
 
-from . import factory_batch, factory_batch_processing, factory_business, factory_completed_filing
+from .. import factory_batch, factory_batch_processing, factory_business, factory_completed_filing
 
 
 RESTORATION_FILING = copy.deepcopy(FILING_HEADER)
 RESTORATION_FILING['filing']['restoration'] = RESTORATION
-
-
-def test_check_run_schedule():
-    """Assert that schedule check validates the day of run based on cron string in config."""
-    with patch.object(Configuration, 'find_by_name') as mock_find_by_name:
-        mock_stage_1_config = MagicMock()
-        mock_stage_2_config = MagicMock()
-        mock_stage_3_config = MagicMock()
-        mock_stage_1_config.val = '0 0 * * 1-2'
-        mock_stage_2_config.val = '0 0 * * 2'
-        mock_stage_3_config.val = '0 0 * * 3'
-        mock_find_by_name.side_effect = [mock_stage_1_config, mock_stage_2_config, mock_stage_3_config]
-
-        with patch('furnish.datetime', wraps=datetime) as mock_datetime:
-            mock_datetime.today.return_value = datetime(2024, 6, 4, 1, 2, 3, 4)
-            cron_valid_1, cron_valid_2, cron_valid_3 = check_run_schedule()
-
-            assert cron_valid_1 is True
-            assert cron_valid_2 is True
-            assert cron_valid_3 is False
 
 
 @pytest.mark.parametrize(
@@ -64,9 +46,9 @@ def test_get_email_address_from_auth(session, test_name, mock_return):
     token = 'token'
     mock_response = MagicMock()
     mock_response.json.return_value = mock_return
-    with patch('furnish.AccountService.get_bearer_token', return_value=token):
-        with patch('furnish.requests.get', return_value=mock_response):
-            email = get_email_address_from_auth('BC1234567')
+    with patch.object(AccountService, 'get_bearer_token', return_value=token):
+        with patch.object(requests, 'get', return_value=mock_response):
+            email = StageOneProcessHelper._get_email_address_from_auth('BC1234567')
             if test_name == 'NO_EMAIL':
                 assert email is None
             else:
@@ -117,9 +99,9 @@ async def test_stage_1_process_first_notification_email(app, session, test_name,
         factory_completed_filing(business, RESTORATION_FILING, filing_type='restoration')
 
     qsm = MagicMock()
-    with patch('furnish.get_email_address_from_auth', return_value=email):
-        with patch('furnish.send_email', return_value=None) as mock_send_email:
-            await stage_1_process(app, qsm)
+    with patch.object(StageOneProcessHelper, '_get_email_address_from_auth', return_value=email):
+        with patch.object(StageOneProcessHelper, '_send_email', return_value=None) as mock_send_email:
+            await process(app, qsm)
 
             if email:
                 mock_send_email.assert_called()
